+++ metadata.rb
@@ -1,7 +1,9 @@
 name             'acmesmith'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+maintainer       'eLitmus Evaluations Pvt Ltd'
+maintainer_email 'crahul@elitmus.com'
 license          'All rights reserved'
 description      'Installs/Configures acmesmith'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
+source_url 		 'https://commit-elitmus.in/crahul/acmesmith-cookbook' if respond_to?(:source_url)
+issues_url       'https://commit-elitmus.in/crahul/acmesmith-cookbook/issues' if respond_to?(:issues_url)